Arreglo pattern CL,DIR. Añadir Precio por Producto

+++ administrador/views.py
@@ -1,9 +1,11 @@
+from datetime import date
 from http import client
 import re
 from django.forms import ValidationError
 from django.http import HttpRequest, HttpResponseRedirect, JsonResponse
 from django.shortcuts import render
 from .models import Clients,Address, Order_Items, Orders, Products, Products_Details
+from django.db.models import Q
 
 # Create your views here.
 def inicio(request):
@@ -36,10 +38,6 @@ def modificarCliente(request):
     Clients.objects.filter(pk=int(id)).update(name=request.POST['nombreClienteModificar'],email=request.POST['emailClienteModificar'],tel=request.POST['telefonoClienteModificar'])
     return HttpResponseRedirect(f'/detallesCliente/{id}')
 
-def borrarCliente(request,id):
-    Clients.objects.get(pk=id).delete()
-    return HttpResponseRedirect('/listarClientes')
-
 def cambiarEstadoCliente(request):
     id=request.POST['idCliente']
     cliente=Clients.objects.filter(pk=int(id))
@@ -47,6 +45,19 @@ def cambiarEstadoCliente(request):
     data = {
        'estado_cambiado': cliente[0].unable
     }
+    return JsonResponse(data)
+
+def comprobarCliente(request):
+    cliente = int(request.POST['idCliente'])
+    if Orders.objects.filter(client=cliente).exists():
+        data = {
+            'tiene_pedido': True
+        }
+    else:
+        Clients.objects.get(pk=cliente).delete()
+        data = {
+            'tiene_pedido': False
+        }
     return JsonResponse(data)
 
 #DIRECCIONES
@@ -68,11 +79,14 @@ def modificarDireccion(request):
     Address.objects.filter(pk=int(idDireccion)).update(street=request.POST['calleDireccionModificar'],city=request.POST['ciudadDireccionModificar'],code=request.POST['codigoDireccionModificar'],country=request.POST['paisDireccionModificar'])
     return HttpResponseRedirect(f'/detallesCliente/{Address.objects.get(pk=int(idDireccion)).client.id}')
 
-def borrarDireccion(request,id):
-    direccion=Address.objects.get(pk=id)
-    cliente=direccion.client.id
-    direccion.delete()
-    return HttpResponseRedirect(f'/detallesCliente/{cliente}')
+def comprobarDireccion(request):
+    direccion=Address.objects.get(pk=request.POST['idDireccion'])
+    if Orders.objects.filter(Q(billing_address=direccion.id) | Q(shipping_address=direccion.id)):
+        data={"tiene_pedido":True}
+    else:
+        data={"tiene_pedido":False}
+        direccion.delete()
+    return JsonResponse(data)
 
 
 #PRODUCTOS
@@ -88,8 +102,8 @@ def crearProductoVista(request):
     return render(request,'crearProducto.html')
 
 def crearProducto(request):
-    Products(name=request.POST['nombreProductoNuevo'],imageURL=request.POST['urlImagenProductoNuevo']).save()
-    producto=Products.objects.get(name=request.POST['nombreProductoNuevo'])
+    producto=Products(name=request.POST['nombreProductoNuevo'],imageURL=request.POST['urlImagenProductoNuevo'])
+    producto.save()
     pesos=[request.POST['pesoProductoNuevoEstandar'],request.POST['pesoProductoNuevoPequenho'],request.POST['pesoProductoNuevoGrande']]
     if pesos[0]=='':
         Products_Details(product=producto,size='AV',price=request.POST['precioProductoNuevoEstandar'],weight=None).save()
@@ -149,18 +163,38 @@ def modificarProducto(request):
             Products_Details.objects.filter(product=producto,size='BG').update(price=request.POST['precioProductoModificarGrande'],weight=float(request.POST['pesoProductoModificarGrande']))
     except:
         pass
-
-
-
     return HttpResponseRedirect(f'/detallesProducto/{id}')
 
-def borrarProducto(request,id):
-    pass
+def borrarProducto(request):
+    producto=Products.objects.get(pk=int(request.POST['idProducto']))
+    tiene=False
+    for item in Products_Details.objects.filter(product=producto.id):
+        if Order_Items.objects.filter(product=item.id).exists():
+            tiene=True
+    if not tiene:
+        data ={
+            'tiene_pedido':False
+        }
+        producto.delete()
+    else:
+        data ={
+            'tiene_pedido':True
+        }
+    return JsonResponse(data)
 
 def comprobarProducto(request):
     producto = request.GET['producto']
     data = {
        'producto_existe':Products.objects.filter(name=producto).exists()
+    }
+    return JsonResponse(data)
+
+def cambiarEstadoProducto(request):
+    id=request.POST['idProducto']
+    producto=Products_Details.objects.filter(pk=int(id))
+    producto.update(unable=not producto[0].unable)
+    data = {
+       'estado_cambiado': producto[0].unable
     }
     return JsonResponse(data)
 
@@ -170,7 +204,11 @@ def listarPedidos(request):
 
 def detallesPedido(request,id):
     pedido=Orders.objects.get(id=int(id))
-    return render(request,'detallesPedido.html',{"pedido":pedido})
+    detalles=Order_Items.objects.filter(order=pedido.id)
+    total=0
+    for item in detalles:
+        total+=(float(item.product.price)*int(item.quantity))
+    return render(request,'detallesPedido.html',{"pedido":pedido,"productos_del_pedido":detalles,"total":total})
 
 def crearPedidoVista(request):
     listaClientes=[]
@@ -182,15 +220,21 @@ def crearPedidoVista(request):
 def crearPedido(request):
     listaProductos=[]
     listaCantidades=[]
+    listaPrecios=[]
     if request.POST.getlist('producto'):
         for item in request.POST.getlist('producto'):
-            listaProductos.append(Products_Details.objects.filter(pk=int(item)))
-        for item in request.POST.getlist('producto'):
-            listaCantidades.append(int(item))
-    Orders(client=request.POST['selectClientes'])
+            listaProductos.append(Products_Details.objects.get(pk=int(item)))
+    if request.POST.getlist('cantidad'):
+        for item in request.POST.getlist('cantidad'):
+            listaCantidades.append(item)
+    if request.POST.getlist('precio'):
+        for item in request.POST.getlist('precio'):
+            listaPrecios.append(item)   
+    pedido=Orders(client=Clients.objects.get(pk=int(request.POST['selectClientes'])),order_date=date.today().strftime("%Y-%m-%d %H:%M:%S"),delivery_date=date.today().strftime("%Y-%m-%d %H:%M:%S"),billing_address=Address.objects.get(pk=int(request.POST['selectFacturacion'])),shipping_address=Address.objects.get(pk=int(request.POST['selectEnvio'])))
+    pedido.save()
     for index in range(len(listaProductos)):
-        Order_Items
-    return HttpResponseRedirect('/listarClientes')
+        Order_Items(order=pedido,product=listaProductos[index],quantity=listaCantidades[index],price=float(listaPrecios[index])).save()
+    return HttpResponseRedirect('/listarPedidos')
 
 def modificarPedidoVista(request,id):
     producto=Products.objects.get(pk=int(id))
@@ -201,8 +245,10 @@ def modificarPedido(request):
     Products.objects.filter(pk=int(id)).update(name=request.POST['nombreProductoModificar'],size=request.POST['tiposTamanhoModificar'],weight=request.POST['pesoProductoModificar'],imageURL=request.POST['urlImagenProductoModificar'],price=request.POST['precioProductoModificar'])
     return HttpResponseRedirect(f'/detallesProducto/{id}')
 
-def borrarPedido(request,id):
-    pass
+def eliminarPedido(request):
+    pedido = int(request.POST['idPedido'])
+    Orders.objects.get(pk=pedido).delete()
+    return JsonResponse({'pedido_eliminado':True})
 
 def direccionesCliente(request):
     idCliente = request.GET['idCliente']
